[DRUP-566] Move the plan controller proxies to `Storage/Controller`.
