+++ apigee_m10n.services.yml
@@ -12,10 +12,10 @@ services:
     arguments: ['@apigee_edge.sdk_connector']
 
   apigee_m10n.sdk_controller_proxy.rate_plan:
-    class: Drupal\apigee_m10n\Entity\Storage\RatePlanSdkControllerProxy
+    class: Drupal\apigee_m10n\Entity\Storage\Controller\RatePlanSdkControllerProxy
 
   apigee_m10n.sdk_controller_proxy.subscription:
-    class: Drupal\apigee_m10n\Entity\Storage\DeveloperAcceptedRatePlanSdkControllerProxy
+    class: Drupal\apigee_m10n\Entity\Storage\Controller\DeveloperAcceptedRatePlanSdkControllerProxy
 
   apigee_m10n.validate_monetization_enabled:
     class: Drupal\apigee_m10n\EventSubscriber\ValidateMonetizationEnabledSubscriber
