+++ src/Entity/Storage/Controller/DeveloperAcceptedRatePlanSdkControllerProxyInterface.php
@@ -17,7 +17,7 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
-namespace Drupal\apigee_m10n\Entity\Storage;
+namespace Drupal\apigee_m10n\Entity\Storage\Controller;
 
 use Apigee\Edge\Entity\EntityInterface;
 use Drupal\apigee_edge\Entity\Controller\EdgeEntityControllerInterface;
