+++ src/Entity/Storage/SubscriptionStorage.php
@@ -21,6 +21,7 @@ namespace Drupal\apigee_m10n\Entity\Storage;
 
 use Drupal\apigee_edge\Entity\Controller\EdgeEntityControllerInterface;
 use Drupal\apigee_edge\Entity\Storage\EdgeEntityStorageBase;
+use Drupal\apigee_m10n\Entity\Storage\Controller\DeveloperAcceptedRatePlanSdkControllerProxyInterface;
 use Drupal\apigee_m10n\Entity\SubscriptionInterface;
 use Drupal\Component\Datetime\TimeInterface;
 use Drupal\Core\Cache\CacheBackendInterface;
